# Conflicts:
#	Ateam_Spring/src/main/java/user/dto/AppUserDTO.java

+++ Ateam_Spring/src/main/java/user/dto/AppUserDTO.java
@@ -1,7 +1,7 @@
 package user.dto;
 
 public class AppUserDTO {
-	private int user_id;
+	private long user_id;
 	private String user_email,
 				user_pw,
 				user_nm,
@@ -14,7 +14,9 @@ public class AppUserDTO {
 	public AppUserDTO() {}
 	
 	//비밀번호를 제외한 모든 정보를 보낼때
+
     public AppUserDTO(int user_id, String user_email, String user_pw, String user_nm, String user_addr, String user_pro_img, String user_phone_no, String user_grade) {
+
         this.user_id = user_id;
         this.user_email = user_email;
         this.user_pw = user_pw;
@@ -36,11 +38,11 @@ public class AppUserDTO {
 		this.user_phone_no = user_phone_no;
 	}
 
-	public int getUser_id() {
+	public long getUser_id() {
 		return user_id;
 	}
 
-	public void setUser_id(int user_id) {
+	public void setUser_id(long user_id) {
 		this.user_id = user_id;
 	}
